refactor(markdown-editor): extract handleWikilinkClick to navigation.service
Context
-------
MarkdownEditor.svelte's `handleEditorClick` was 90 lines of wikilink
navigation orchestration inlined in a UI component: file-tree
flattening, wikilink resolution, periodic-note detection,
non-existent target creation, anchor position lookup, tab
open/switch, pending-scroll bookkeeping. The component had 12
imports from 8 unrelated areas of the app (periodic-notes plugin,
backlinks, filesystem, quick-switcher, etc.) just to serve this one
handler. That violates the usual split where a Svelte component
stays visual and delegates orchestration to a service.

Solution
--------
Move the orchestration into a new
`extensions/wikilink/navigation.service.ts ›
handleWikilinkClick(view, target)`. Also pull out the repeated
"heading-or-blockId → position" lookup into a small local
`resolveAnchorPosition(content, info)` helper — it was written out
three times in the original handler.

MarkdownEditor.svelte's `handleEditorClick` becomes 11 lines: early-
return for non-wikilink clicks, preventDefault + stopPropagation,
delegate to the service. Nine imports removed from the component:
findWikilinkInfoAtPosition, findHeadingPosition, findBlockIdPosition,
openFileInEditor, resolveWikilink, flattenFileTree, fsStore,
createFile, vaultStore, detectPeriodicNoteType,
openOrCreatePeriodicNoteForDate.

Testing
-------
Full suite: 5057 tests green. pnpm check clean.

Files
-----
- src/lib/core/markdown-editor/extensions/wikilink/navigation.service.ts
  New (120 lines). Single exported `handleWikilinkClick(view, target)`
  returning `Promise<boolean>` (true when the click was a wikilink
  and was handled). Private `resolveAnchorPosition` helper folds the
  heading/block-id branching the old handler repeated three times.
- src/lib/core/markdown-editor/MarkdownEditor.svelte
  (~11, ~44-55): drop 9 imports, replace 87-line handleEditorClick
  with an 11-line shim that delegates to the service.
